replace sizet_t by  uint , use  set STL

+++ projectSampling/include/dmatrix.h
@@ -29,18 +29,18 @@ struct SVector {
 class DMatrix{
 public:    
     DMatrix();
-    DMatrix(size_t r_dim, size_t c_dim);
-    DMatrix(size_t r_dim, size_t c_dim, double *ptr);
+    DMatrix(uint r_dim, uint c_dim);
+    DMatrix(uint r_dim, uint c_dim, double *ptr);
 	~DMatrix();
-    double& operator()(size_t i, size_t j);
-	double operator()(size_t i, size_t j)const;
-	size_t randRow(size_t n, double sum);
-	size_t randCol(size_t m, double sum);
+    double& operator()(uint i, uint j);
+	double operator()(uint i, uint j)const;
+	uint randRow(uint n, double sum);
+	uint randCol(uint m, double sum);
     void init(uint TYPE, double mean, double stdev);
     double* get_abs_row_sum();
     double* get_abs_col_sum();
 public:
-	size_t row, col;
+	uint row, col;
 	double *value;
 private:
     bool SELFINITVALUE;
@@ -50,14 +50,14 @@ private:
 class DVector{
 public:
     DVector();
-    DVector(size_t v_dim);
-    DVector(size_t v_dim, double*ptr);
+    DVector(uint v_dim);
+    DVector(uint v_dim, double*ptr);
     void init(uint TYPE, double mean, double stdev);
 	~DVector();
-    double& operator()(size_t n);
-	double operator()(size_t n)const;
+    double& operator()(uint n);
+	double operator()(uint n)const;
 public:
-	size_t dim;
+	uint dim;
 	double* value;
 private:
     bool SELFINITVALUE;
@@ -75,13 +75,13 @@ DMatrix::DMatrix(){
     value = nullptr;
     SELFINITVALUE = false;
 }
-DMatrix::DMatrix(size_t r_dim, size_t c_dim){
+DMatrix::DMatrix(uint r_dim, uint c_dim){
     row = r_dim;
     col = c_dim;
     value = nullptr;
     SELFINITVALUE = false;
 }
-DMatrix::DMatrix(size_t r_dim, size_t c_dim, double *ptr){
+DMatrix::DMatrix(uint r_dim, uint c_dim, double *ptr){
     row = r_dim;
     col = c_dim;
     value = ptr;
@@ -94,12 +94,12 @@ void DMatrix::init(uint TYPE, double mean, double stdev){
     }
     switch (TYPE) {
         case INIT_RAND_N:
-            for(size_t i = 0; i < col*row; ++i){
+            for(uint i = 0; i < col*row; ++i){
                 // normal distribution with \mu = mean,\sigma^2 = stdev
                 value[i] = rnd::randn(mean, stdev);
             }break;
         case INIT_RAND_U:
-            for(size_t i = 0; i < col*row; ++i)
+            for(uint i = 0; i < col*row; ++i)
                 value[i] = rnd::randu();
             break;
         case INIT_ALL_ONES:
@@ -111,18 +111,18 @@ void DMatrix::init(uint TYPE, double mean, double stdev){
         default: break;
     }
 }
-double& DMatrix::operator()(size_t i, size_t j){ 
+double& DMatrix::operator()(uint i, uint j){ 
     //assert((i >= row) && (j >= col));
     return value[j*row +i];
 }
-double DMatrix::operator()(size_t i, size_t j) const {
+double DMatrix::operator()(uint i, uint j) const {
     return value[j*row +i];
 }
-size_t DMatrix::randRow(size_t n, double sum){
+uint DMatrix::randRow(uint n, double sum){
     double x,temp;
     x = sum*((double)rand()/(double)RAND_MAX);
     temp = 0;
-    for (size_t i = 0; i < row; ++i){
+    for (uint i = 0; i < row; ++i){
         temp += abs(value[i + n*row]);
         if(x <= temp){ 
             return i;
@@ -130,11 +130,11 @@ size_t DMatrix::randRow(size_t n, double sum){
     }
     return (row-1);
 }
-size_t DMatrix::randCol(size_t m, double sum){
+uint DMatrix::randCol(uint m, double sum){
     double x,temp;
     x = sum*((double)rand()/(double)RAND_MAX);
     temp = 0;
-    for (size_t j = 0; j < col; ++j){
+    for (uint j = 0; j < col; ++j){
         temp += abs(value[m + j*row]);
         if(x <= temp){ 
             return j;
@@ -148,21 +148,21 @@ DVector::DVector(){
     value = nullptr;
     SELFINITVALUE = false;
 }
-DVector::DVector(size_t v_dim){
+DVector::DVector(uint v_dim){
     dim = v_dim;
     value = nullptr;
     SELFINITVALUE = false;
 }
-DVector::DVector(size_t v_dim, double *ptr){
+DVector::DVector(uint v_dim, double *ptr){
     dim = v_dim;
     value = ptr;
     SELFINITVALUE = false;
 }
-double& DVector::operator()(size_t n){ 
+double& DVector::operator()(uint n){ 
     //assert((i >= row) && (j >= col));
     return value[n];
 }
-double DVector::operator()(size_t n) const {
+double DVector::operator()(uint n) const {
     return value[n];
 }
 
@@ -173,12 +173,12 @@ void DVector::init(uint TYPE, double mean, double stdev){
     }
     switch (TYPE) {
         case INIT_RAND_N:
-            for(size_t i = 0; i < dim; ++i)
+            for(uint i = 0; i < dim; ++i)
                 // normal distribution with \mu = mean,\sigma^2 = stdev
                 value[i] = rnd::randn(mean, stdev);
             break;
         case INIT_RAND_U:
-            for(size_t i = 0; i < dim; ++i)
+            for(uint i = 0; i < dim; ++i)
                 value[i] = rnd::randu();
             break;
         case INIT_ALL_ONES:
